Test examples for json Parameters
Zermelo example modified to use Json parametrs
Fix of iterations stopping threshold used in main.
The isse was related to the nulber of deleted points used to stom the iterations.
If the initial grid is not fine ( in case of refinments)
and the stopping number is important ( 50) then the algorithm
stops with bad quality of kernel computed.
To solve this a nex  algo parameter is  introduced
to allow to modify the number of points used
as stopping threshold
from json file