Improving optimal trajectory reconstruction
Thsi dev improves the trajectory reconstruction method
for tiem-optima problems.
At each step we first compute teh optimal control and successor on the
grid Grid. These are guaranteed as the viability algorithm computes an approximation
of teh viable set on the give grid.
Then we try to compute a control and such that the trajectory approximation
starting from a point that is not on a grid is the nearest to the discrete approximation.