Integration of spdlog
This dev integrates apd log in ViabLab to
add a logging feature, simple to use in the code.
Some clean up of the code , replacing multiple console outputs by
 structered logging messages. Logging is done into rwo sources with different
levels : a file with minimu level = trace
to get all the execution details. And console with minimum level = info
to show more light oututs during the execution of the program.