Rework of Control objects
This change adds a ControlGrid class to simplyfy the management
of control grid parameters. This will clean up the SysDyn class code.
This clss will simply point to different ControlGrid objects depending on
the type of dynamics.
Also this change add some rework of ViabilityTrajectory helpers, with a
common parent class and grid specific implementations.
Some methods that was previousely added to SysDyn calss are moved to
ViabiTrajectoryHelper as it is more the responsability of this class.
This change impacts controlPickers and ControlPickStrategies and
simplifies their dependencies.